Move followed implementation out of shows module
