--- data/profile/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/profile/api/followed/FollowedRepository.kt
@@ -1,14 +1,0 @@
-package com.thomaskioko.tvmaniac.profile.api.followed
-
-import com.thomaskioko.tvmaniac.core.networkutil.Either
-import com.thomaskioko.tvmaniac.core.networkutil.Failure
-import kotlinx.coroutines.flow.Flow
-
-//TODO:: Move to separate module. #82
-interface FollowedRepository {
-
-    fun observeUpdateFollowedShow(traktId: Long, addToWatchList: Boolean): Flow<Either<Failure, Unit>>
-
-    suspend fun syncFollowedShows()
-
-}
